Huge refactor to eliminate circular dependencies
