+++ src/utils/paperColour.ts
@@ -11,11 +11,11 @@ import { DEFAULT_PAPER_COLOUR, DARK_PAPER_COLOUR, LIGHT_PAPER_COLOUR } from "../
 import { PAGE_NODE_ATTR_KEYS } from "../constants/page";
 import { DARK_THEME } from "../constants/theme";
 import { Nullable } from "../types/record";
-import { getPageAttributeByPageNum, isPageNode } from "./page";
+import { getPageAttributeByPageNum, isPageNode } from "./nodes/page/page";
 import { isValidColour } from "./colour";
 import { nodeHasAttribute } from "./node";
 import { getDeviceTheme } from "./theme";
-import { setPageNodeAttribute } from "./setPageAttributes";
+import { setPageNodeAttribute } from "./nodes/page/setPageAttributes";
 
 /**
  * Get the paper colour based on the device theme.
