+++ src/utils/setSideConfig.ts
@@ -10,8 +10,8 @@ import { Dispatch } from "@tiptap/core";
 import { pageSides } from "../constants/pageSides";
 import { PageSide, MultiSide, MultiAxisSide } from "../types/page";
 import { Nullable } from "../types/record";
-import { getPageNodePosByPageNum, isPageNode } from "./page";
-import { setPageNodeAttribute } from "./setPageAttributes";
+import { getPageNodePosByPageNum, isPageNode } from "./nodes/page/page";
+import { setPageNodeAttribute } from "./nodes/page/setPageAttributes";
 import { NodePos } from "../types/node";
 
 type SideConfig<V> = { [key in PageSide]: V };
